requirements.txt
adafruit-io==1.1.0
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
paho-mqtt==1.3.1
PyYAML==3.12
requests==2.18.4
urllib3==1.22
yoctopuce==1.10.28878
